+++ lib/application.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:vision_app/core/flavor/flavor_config.dart';
 import 'package:vision_app/core/ui/theme/vision_theme.dart';
 import 'package:vision_app/core/utils/constants/routes/vision_routes.dart';
+import 'package:vision_app/features/splash/page/splash_page.dart';
 import 'package:vision_app/init_depedencies.dart';
 
 void initializeApp(FlavorConfig flavor) {
@@ -18,53 +19,8 @@ class MyApp extends StatelessWidget {
       title: 'Vision App',
       debugShowCheckedModeBanner: false,
       theme: VisionTheme.darkTheme,
-      home: const MyHomePage(title: 'Vision App Home Page'),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key, required this.title});
-
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      _counter++;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        title: Text(widget.title),
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            const Text('You have pushed the button this many times:'),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headlineMedium,
-            ),
-          ],
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: const Icon(Icons.add),
-      ),
+      initialRoute: VisionRoutes.initial,
+      routes: {VisionRoutes.initial: (context) => SplashPage()},
     );
   }
 }
